requirements.txt
streamlit==1.49.1
langgraph==0.6.6
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.2.8
pandas==2.3.2
plotly==6.3.0
altair==5.5.0
matplotlib==3.10.6
requests==2.32.5
python-dotenv==1.1.1
openai==1.106.1
numpy==2.3.2
kaggle==1.5.16
